requirements.txt
django
djangorestframework
jsonfield